+++ source/Core/Field.php
@@ -8,146 +8,91 @@
 namespace OxidEsales\EshopCommunity\Core;
 
 use OxidEsales\Eshop\Core\Str;
+use OxidEsales\EshopCommunity\Internal\Container\ContainerFactory;
 
 use function is_string;
 
-/**
- * Database field description object.
- *
- * when a value is requested from oxField, it either takes 'value' field or
- * 'rawValue' (depending which one is available).
- * In case 'rawValue' is taken, it is escaped first before returning.
- *
- * T_RAW and T_TEXT types represent not the assignment logic, but rather a
- * returned value escaping status.
- */
-class Field // extends \OxidEsales\Eshop\Core\Base
+class Field
 {
     /**
      * escaping functionality type: expected value is escaped text.
      */
-    const T_TEXT = 1;
+    public const T_TEXT = 1;
 
     /**
      * escaping functionality type: expected value is not escaped (raw) text.
      */
-    const T_RAW = 2;
+    public const T_RAW = 2;
 
-    /**
-     * Constructor
-     * Initial value assigment is coded here by not calling a function is for performance
-     * because oxField is created MANY times and even a function call matters
-     *
-     * if T_RAW is used, then it fills $value, because this is the value, that does
-     * not need to be escaped and is by definition equal to $rawValue (which is not set
-     * for less memory usage).
-     *
-     * if T_TEXT is used, then $rawValue is assigned and retrieved $value (which is not
-     * set initially) is escaped $rawValue.
-     *
-     * e.g.
-     * > if your input is "<b>string</b>" and you want your output to be exactly same,
-     *   you should use T_RAW - in this way it will be assigned to $value property as
-     *   the result.
-     * > if your input is "1 & (a < b)" and you want your output to be escaped, you
-     *   should use T_TEXT - in this way it will be assigned to $rawValue property and
-     *   it will be escaped.
-     *
-     * @param mixed $value Field value
-     * @param int   $type  Value type
-     *
-     * @return null
-     */
     public function __construct($value = null, $type = self::T_TEXT)
     {
-        // duplicate content here is needed for performance.
-        // as this function is called *many* (a lot) times, it is crucial to be fast here!
         $this->rawValue = $value;
-        if ($type == self::T_RAW) {
+        if ((int)$type === self::T_RAW) {
             $this->value = $value;
         }
     }
 
-    /**
-     * Checks if $name is set
-     *
-     * @param string $name Variable name
-     *
-     * @return boolean
-     */
-    public function __isset($name)
+    public function __isset($name): bool
     {
         return $this->{$name} !== null;
     }
 
     /**
-     * Magic getter
-     *
-     * @param string $name Variable name
-     *
-     * @return string|null
+     * @param string $name
+     * @return mixed|string|null
      */
-    public function __get($name)
+    public function __get(string $name)
     {
-        switch ($name) {
-            case 'rawValue':
-                return $this->value;
-                break;
-            case 'value':
-                if (is_string($this->rawValue)) {
-                    $this->value = Str::getStr()->htmlspecialchars($this->rawValue);
-                } else {
-                    // TODO: call htmlentities for each value (recursively?)
-                    $this->value = $this->rawValue;
-                }
-                if ($this->rawValue == $this->value) {
+        if (!($name === 'value' || $name === 'rawValue')) {
+            return null;
+        }
+        if ($name === 'value') {
+            if ($this->valueNeedsEscaping()) {
+                $escapedValue = Str::getStr()->htmlspecialchars($this->rawValue);
+                $this->value = $escapedValue;
+                if ($escapedValue === (string)$this->rawValue) {
                     unset($this->rawValue);
                 }
-                return $this->value;
-                break;
-            default:
-                return null;
-                break;
+            } else {
+                $this->value = $this->rawValue;
+                unset($this->rawValue);
+            }
         }
+        return $this->value;
+    }
+
+    public function __toString(): string
+    {
+        return (string)$this->value;
     }
 
     /**
-     * Returns actual field value
-     *
-     * @return string
+     * @param $value
+     * @param $type
+     * @return void
      */
-    public function __toString()
+    public function setValue($value = null, $type = self::T_TEXT): void
     {
-        return (string) $this->value;
+        unset($this->rawValue, $this->value);
+        $this->initValue($value, $type);
     }
 
     /**
-     * Converts to formatted db date
-     * @deprecated method will be removed in v7.0. Use \OxidEsales\EshopCommunity\Core\UtilsDate::formatDBDate() directly.
+     * @return mixed
      */
-    public function convertToFormattedDbDate()
+    public function getRawValue(): mixed
     {
-        $this->setValue(\OxidEsales\Eshop\Core\Registry::getUtilsDate()->formatDBDate($this->rawValue), self::T_RAW);
+        return $this->rawValue ?? $this->value;
     }
 
     /**
-     * Converts to pseudo html - new lines to <br /> tags
-     * @deprecated method will be removed in v7.0. Use htmlspecialchars() abd nl2br() directly.
+     * @param $value
+     * @param $type
+     * @return void
      */
-    public function convertToPseudoHtml()
+    protected function initValue($value = null, $type = self::T_TEXT): void
     {
-        $this->setValue(str_replace("\r", '', nl2br(Str::getStr()->htmlspecialchars($this->rawValue))), self::T_RAW);
-    }
-
-    /**
-     * Initial field value
-     *
-     * @param mixed $value Field value
-     * @param int   $type  Value type
-     */
-    protected function initValue($value = null, $type = self::T_TEXT)
-    {
-        if ($type == self::T_TEXT) {
+        if ((int)$type === self::T_TEXT) {
             $this->rawValue = $value;
         } else {
             $this->value = $value;
@@ -155,29 +100,13 @@ class Field // extends \OxidEsales\Eshop\Core\Base
     }
 
     /**
-     * Sets field value and type
-     *
-     * @param mixed $value Field value
-     * @param int   $type  Value type
+     * @return bool
      */
-    public function setValue($value = null, $type = self::T_TEXT)
+    private function valueNeedsEscaping(): bool
     {
-        unset($this->rawValue);
-        unset($this->value);
-        $this->initValue($value, $type);
-    }
-
-    /**
-     * Return raw value
-     *
-     * @return string
-     */
-    public function getRawValue()
-    {
-        if (null === $this->rawValue) {
-            return $this->value;
-        };
-
-        return $this->rawValue;
+        return is_string($this->rawValue)
+            && !ContainerFactory::getInstance()
+                ->getContainer()
+                ->getParameter('oxid_esales.templating.engine_autoescapes_html');
     }
 }